fix(oauth): PropertiesSanitizer should not filter out subject and actor tokens
As a general rule, the sanitizer should filter out credentials such as: usernames, passwords, secrets, private keys, etc.

But tokens are generally tolerated. In this regard, it is incorrect for the sanitizer to filter out subject and actor tokens.

This PR also introduces unit tests for both LegacyPropertiesMigrator and PropertiesSanitizer.