fix(tests): Refactor user emulation code
Fixes #97.

This change fixes various race conditions in tests:

* In the user emulator classes;
* In expectation classes (for Authorization Code and Device Code);
* In flow classes (for Authorization Code and Device Code).

One of the reported race conditions is #97, which is due to 2 concurrent agents asking for authentication in parallel.

The `UserEmulator` API was completely redesigned (and considerably simplified).

The affected expectation classes were refactored to remove internal fields that were causing race conditions. These were replaced with maps that keep track of all the in-flight requests.

The affected flow classes now synchronize on the console PrintStream before printing messages, in order to avoid garbled messages when two agents attempt to print at the same time.