feat(tests): Test against a matrix of Iceberg and Spark/Flink versions
This change introduces a "simple" way to play Spark and Flink tests against a matrix of Iceberg and Spark/Flink versions.

It is "simple" because it may not work if different engine versions become so incompatible, that the test classes themselves need to be changed. Once that happens (and it may soon, with Spark 4.0) then we'll need to look for a more complex setup.

But for now, this change at least allows testing against Iceberg 1.9.0 and 1.9.1, and against the two last versions of each engine.

The versions to use are controlled by the following properties:

* `authmgr.test.iceberg.versions`: 1.9.1, 1.9.0
* `authmgr.test.spark.versions`: 3.5.6, 3.5.5
* `authmgr.test.flink.versions`: 1.20.2, 1.20.1

Summary of tasks:

* The `intTest` task runs the tests with the latest versions.
* The `intTestMatrix` task runs the entire matrix.
* The `intTest_iceberg190_spark355` task runs the tests with Iceberg 1.9.0 and Spark 3.5.5. There are similar tasks for each combination of iceberg/engine versions available.
* The `printTestMatrix` task prints the different tasks available.